Issue/262 shadows flicker (#263)
* #237 Flip a bit on shadow-impacted frames, use this to govern stencil write.

* Attempts to eliminate shadow flicker (which works), but introduces strange crash.

* Allow users to disable threads with threads -1.

* Fix multithreading crash in shadow calculation by using a dynamically allocated array of shadow counts.

* Remove r_bsp_plane_t typedef, this will bring is closer to the old code, which is probably for the best.

* Revert some of r_bsp.c so that it looks as it used to.

* Ensure sustained lights are added before we cull entities.

This allows sustained light sources to contribute to illuminations and
shadows. Looks so much better (again). I dunno how we missed this for
so long.

* Use an array NOT within r_bsp_model_t or its members to track shadow counts

This alleviates crashes with multithreaded clients. The compiler must
be doing some serious optimizations around const.

* Xchode.

+++ src/thread.h
@@ -50,7 +50,7 @@ thread_t *Thread_Create_(const char *name, ThreadRunFunc run, void *data);
 #define Thread_Create(f, d) Thread_Create_(#f, f, d)
 void Thread_Wait(thread_t *t);
 uint16_t Thread_Count(void);
-void Thread_Init(uint16_t num_threads);
+void Thread_Init(ssize_t num_threads);
 void Thread_Shutdown(void);
 
 #endif /*__THREAD_H__ */
